+++ pkg/apis/deployment/v1/server_group_spec.go
@@ -23,6 +23,7 @@ package v1
 import (
 	"math"
 	"strings"
+	"time"
 
 	"github.com/arangodb/kube-arangodb/pkg/util/errors"
 
@@ -149,6 +150,8 @@ type ServerGroupSpec struct {
 	InternalPort *int `json:"internalPort,omitempty"`
 	// AllowMemberRecreation allows to recreate member. Value is used only for Coordinator and DBServer with default to True, for all other groups set to false.
 	AllowMemberRecreation *bool `json:"allowMemberRecreation,omitempty"`
+	// TerminationGracePeriodSeconds override default TerminationGracePeriodSeconds for pods - via silent rotation
+	TerminationGracePeriodSeconds *int64 `json:"terminationGracePeriodSeconds,omitempty"`
 }
 
 // ServerGroupSpecSecurityContext contains specification for pod security context
@@ -701,4 +704,13 @@ func (s ServerGroupSpec) GetShutdownDelay(group ServerGroup) int {
 		}
 	}
 	return *s.ShutdownDelay
+}
+
+// GetTerminationGracePeriod returns termination grace period as Duration
+func (s ServerGroupSpec) GetTerminationGracePeriod(group ServerGroup) time.Duration {
+	if v := s.TerminationGracePeriodSeconds; v == nil {
+		return group.DefaultTerminationGracePeriod()
+	} else {
+		return time.Second * time.Duration(*v)
+	}
 }